+++ Topos.AzureBlobs.Tests/AzureBlobConfig.cs
@@ -1,8 +1,7 @@
-﻿using Microsoft.Azure.Storage;
-
+﻿
 namespace Topos.AzureBlobs.Tests;
 
 static class AzureBlobConfig
 {
-    public static CloudStorageAccount StorageAccount => CloudStorageAccount.DevelopmentStorageAccount;
+    public const string ConnectionString = "UseDevelopmentStorage=true";
 }