+++ Topos.AzureBlobs/Config/AzureBlobsPositionManagerConfigurationExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Microsoft.Azure.Storage;
 using Topos.AzureBlobs;
 using Topos.Consumer;
 // ReSharper disable UnusedMember.Global
@@ -11,14 +10,14 @@ public static class AzureBlobsPositionManagerConfigurationExtensions
     /// <summary>
     /// Configures Topos to stores its consumer positions in blobs. Individual blobs will be created for each relevant topic/partition.
     /// </summary>
-    public static void StoreInAzureBlobs(this StandardConfigurer<IPositionManager> configurer, CloudStorageAccount storageAccount, string containerName)
+    public static void StoreInAzureBlobs(this StandardConfigurer<IPositionManager> configurer, string connectionString, string containerName)
     {
         if (configurer == null) throw new ArgumentNullException(nameof(configurer));
-        if (storageAccount == null) throw new ArgumentNullException(nameof(storageAccount));
+        if (connectionString == null) throw new ArgumentNullException(nameof(connectionString));
         if (containerName == null) throw new ArgumentNullException(nameof(containerName));
 
         var registrar = StandardConfigurer.Open(configurer);
 
-        registrar.Register(_ => new AzureBlobsPositionManager(storageAccount, containerName));
+        registrar.Register(_ => new AzureBlobsPositionManager(connectionString, containerName));
     }
 }